refac: write modular loading functions
Via specific decorators that focus on one aspect, I was able to break
apart most of the loading functions such that only one line is necessary
to e.g. create a `lymph` model from a parameter YAML file. And these
utilities work both when executed as part of a CLI, as well as in
`streamlit`. In both cases they also report what happens and handle errors
gracefully.